Organize wallet states more
