+++ packages/wallet/src/redux/initializing/reducer.ts
@@ -6,12 +6,11 @@ import {
   WAIT_FOR_ADJUDICATOR,
   WaitForLogin,
   WaitForAdjudicator,
-  waitingForChannelInitialization,
-  InitializedState,
 } from './state';
+import { waitingForChannelInitialization, InitializedState } from '../state';
 
-import { WalletAction, LOGGED_IN, ADJUDICATOR_KNOWN } from './actions';
-import { unreachable } from '../utils/reducer-utils';
+import { WalletAction, LOGGED_IN, ADJUDICATOR_KNOWN } from '../actions';
+import { unreachable } from '../../utils/reducer-utils';
 import { initializationSuccess } from 'magmo-wallet-client/lib/wallet-events';
 
 export const initializingReducer = (
